refactor(website): rename trail-cli to trails-cli
The framework and its CLI are `trails`; the browser CLI adapter, its deps
interface and its factory were spelled singular. Mechanical rename of
`trail-cli.ts` -> `trails-cli.ts`, `TrailCliDeps` -> `TrailsCliDeps` and
`createTrailCLI` -> `createTrailsCLI`, plus the importers and the three
frontiers docs that name the file. No behavior change; the one test name that
mentions the file is left verbatim.
