+++ packages/website/src/lib/frontiers/components/tutorial/StepContent.test.ts
@@ -9,7 +9,7 @@ import initSqlJs, { type SqlJsStatic } from "sql.js";
 import { SqlJsAdapter } from "../../sql-js-adapter.js";
 import { VirtualFS } from "../../virtual-fs.js";
 import type { TutorialStep } from "../../tutorials/types.js";
-import type { CliResult } from "../../trail-cli.js";
+import type { CliResult } from "../../trails-cli.js";
 
 let SQL: SqlJsStatic;
 let adapter: SqlJsAdapter;
